⚡️ Split `flax.py` in `load.py` and `save.py`
As it works with JAX and Haiku too, not just Flax. So we're extending `safejax` use-cases.
