+++ src/safejax/save.py
@@ -1,15 +1,11 @@
-import os
-from pathlib import Path
 from typing import Any, Dict, Union
 
 import numpy as np
-from flax.core.frozen_dict import FrozenDict, freeze
+from flax.core.frozen_dict import FrozenDict
 from jax import numpy as jnp
-from safetensors.flax import load, load_file, save, save_file
+from safetensors.flax import save, save_file
 
 from safejax.typing import PathLike
-
-__all__ = ["serialize", "deserialize"]
 
 
 def flatten_dict(
@@ -33,15 +29,15 @@ def flatten_dict(
     Returns:
         A flattened dictionary containing the model parameters.
     """
-    weights = {}
+    flattened_params = {}
     for key, value in params.items():
         key = f"{key_prefix}.{key}" if key_prefix else key
         if isinstance(value, jnp.DeviceArray) or isinstance(value, np.ndarray):
-            weights[key] = value
+            flattened_params[key] = value
             continue
         if isinstance(value, FrozenDict) or isinstance(value, Dict):
-            weights.update(flatten_dict(params=value, key_prefix=key))
-    return weights
+            flattened_params.update(flatten_dict(params=value, key_prefix=key))
+    return flattened_params
 
 
 def serialize(
@@ -67,45 +63,3 @@ def serialize(
     else:
         save_file(tensors=flattened_dict, filename=filename)
         return filename
-
-
-def unflatten_dict(tensors: Dict[str, jnp.DeviceArray]) -> FrozenDict:
-    """
-    Unflatten a `FrozenDict` from a `Dict` of tensors.
-
-    Reference at https://stackoverflow.com/a/63545677.
-
-    Args:
-        tensors: A `Dict` of tensors containing the model parameters.
-
-    Returns:
-        A `FrozenDict` containing the model parameters.
-    """
-    weights = {}
-    for key, value in tensors.items():
-        subkeys = key.split(".")
-        for subkey in subkeys[:-1]:
-            weights = weights.setdefault(subkey, {})
-        weights[subkeys[-1]] = value
-    return freeze(weights)
-
-
-def deserialize(path_or_buf: Union[PathLike, bytes]) -> FrozenDict:
-    """
-    Deserialize a Flax model from either a `bytes` object or a file path.
-
-    Args:
-        path_or_buf: A `bytes` object or a file path containing the serialized model.
-
-    Returns:
-        A `FrozenDict` containing the model parameters.
-    """
-    if isinstance(path_or_buf, bytes):
-        loaded_dict = load(data=path_or_buf)
-    if (
-        isinstance(path_or_buf, str)
-        or isinstance(path_or_buf, Path)
-        or isinstance(path_or_buf, os.PathLike)
-    ):
-        loaded_dict = load_file(filename=path_or_buf)
-    return unflatten_dict(tensors=loaded_dict)
